+++ packages/falcon-front-kit/src/index.ts
@@ -17,4 +17,5 @@ export * from './Routes';
 export * from './ScrollToTop';
 export * from './Search';
 export * from './SortOrder';
+export * from './SSR';
 export * from './Wizards';